aircraft_manufacturing/inventory/tests.py
from django.test import TestCase
from django.contrib.auth.models import User
from rest_framework.test import APITestCase
from rest_framework import status
from accounts.models import Team, TeamMember
from .models import Part