pyproject.toml
[project]
name = "aircraft-manufacturing"
version = "1.0.0"
description = "Aircraft manufacturing application demo"
readme = "README.md"
requires-python = ">=3.12.7"
dependencies = [
    "django==5.1.5",
    "django-cors-headers==4.6.0",
    "django-datatables-view==1.20.0",
    "django-filter==24.3",
    "django-tables2==2.7.5",
    "djangorestframework==3.15.2",
    "drf-yasg==1.21.8",
    "gunicorn==23.0.0",
    "psycopg2-binary==2.9.10",
    "pytest==8.3.4",
    "pytest-django==4.9.0",
    "python-dotenv==1.0.1",
]
